refactor: 飲み物とラベルの追加・編集をコンポーネント化した
